Add components to files
